telegram_requirements.txt
# Telegram Bot Requirements
# Add these to your existing requirements.txt or install separately

# Telegram Bot Library
python-telegram-bot[all]==20.7
aiohttp==3.9.1
asyncio-mqtt==0.13.0